+++ code/packages/sdk/src/plugins/stage-grid/stage-grid.ts
@@ -12,7 +12,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
   constructor(params: WeaveStageGridPluginParams) {
     super();
 
-    const { gridSize = 50 } = params;
+    const { gridSize = 100 } = params;
 
     this.gridSize = gridSize;
   }
@@ -32,7 +32,10 @@ export class WeaveStageGridPlugin extends WeavePlugin {
   initLayer() {
     const stage = this.instance.getStage();
 
-    const layer = new Konva.Layer({ id: this.getLayerName() });
+    const layer = new Konva.Layer({
+      id: this.getLayerName(),
+      listening: false,
+    });
     stage.add(layer);
   }
 
@@ -44,7 +47,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
   private initEvents() {
     const stage = this.instance.getStage();
 
-    stage.on('mousemove', (e) => {
+    stage.on('mousemove touchmove', (e) => {
       e.evt.preventDefault();
       this.render();
     });
@@ -98,7 +101,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
 
     const stage = this.instance.getStage();
 
-    const size = stage.width() / this.gridSize;
+    const size = stage.width() / (this.gridSize * stage.scaleX());
 
     const delta = 2 * size;
 
@@ -120,6 +123,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
     for (let row = 0; row <= numRows; row++) {
       const pageY = row * size + startPageY;
       const canvasY = pageY - 2 * startPageY;
+
       layer.add(
         new Line({
           points: [
@@ -128,8 +132,9 @@ export class WeaveStageGridPlugin extends WeavePlugin {
             (stage.width() - stage.x() + 2 * delta) / stage.scaleX(),
             canvasY,
           ],
-          stroke: 'rgba(0, 0, 0, 0.2)',
-          strokeWidth: 0.5,
+          stroke: '#cccccc',
+          strokeWidth: 1.5 / stage.scaleX(),
+          listening: false,
         })
       );
     }
@@ -137,6 +142,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
     for (let col = 0; col <= numCols; col++) {
       const pageX = col * size + startPageX;
       const canvasX = pageX - 2 * startPageX;
+
       layer.add(
         new Line({
           points: [
@@ -145,8 +151,9 @@ export class WeaveStageGridPlugin extends WeavePlugin {
             canvasX,
             (stage.height() - stage.y() + 2 * delta) / stage.scaleY(),
           ],
-          stroke: 'rgba(0, 0, 0, 0.2)',
-          strokeWidth: 0.5,
+          stroke: '#cccccc',
+          strokeWidth: 1.5 / stage.scaleX(),
+          listening: false,
         })
       );
     }
@@ -161,7 +168,7 @@ export class WeaveStageGridPlugin extends WeavePlugin {
 
     const stage = this.instance.getStage();
 
-    const size = stage.width() / this.gridSize;
+    const size = stage.width() / (this.gridSize * stage.scaleX());
 
     const delta = 2 * size;
 
@@ -192,10 +199,11 @@ export class WeaveStageGridPlugin extends WeavePlugin {
           new Circle({
             x: canvasX,
             y: canvasY,
-            radius: 1,
-            fill: 'rgba(0, 0, 0, 0.2)',
-            stroke: 'rgba(0, 0, 0, 0.2)',
+            radius: 1.5 / stage.scaleX(),
+            fill: '#cccccc',
+            stroke: '#cccccc',
             strokeWidth: 0,
+            listening: false,
           })
         );
       }
@@ -209,11 +217,13 @@ export class WeaveStageGridPlugin extends WeavePlugin {
   enable() {
     this.enabled = true;
     this.getLayer()?.show();
+    this.render();
   }
 
   disable() {
     this.enabled = false;
     this.getLayer()?.hide();
+    this.render();
   }
 
   getType() {